Rename back to snake_egg and update typings
