+++ Cargo.toml
@@ -4,7 +4,7 @@ version = "0.1.0"
 edition = "2021"
 
 [package.metadata.maturin]
-name = "egg._internal"
+name = "snake_egg._internal"
 
 [lib]
 name = "snake_egg"
