+++ pyproject.toml
@@ -1,13 +1,10 @@
-[project]
-name = "snake-egg"
-
 [build-system]
 requires = ["maturin>=0.11,<0.12"]
 build-backend = "maturin"
 dependencies = ["typing-extensions"]
 
 [project]
-name = "egg"
+name = "snake-egg"
 
 [tool.isort]
 profile = "black"
@@ -20,7 +17,7 @@ dev = [
   "mypy"
 ]
 [tool.mypy]
-ignore_missing_imports = false
+ignore_missing_imports = true
 warn_redundant_casts = true
 check_untyped_defs = true
 strict_equality = true
