+++ snake_egg/snake_egg.pyi
