+++ snake_egg/tests/test_math.py
@@ -10,9 +10,7 @@
 from collections import namedtuple
 from typing import Any, List
 
-from snake_egg import EGraph, Rewrite, vars
-
-from egg import EGraph, Rewrite, vars
+from snake_egg.snake_egg import EGraph, Rewrite, vars
 
 # Operations
 Diff = namedtuple("Diff", "x y")
