+++ snake_egg/tests/test_prop.py
@@ -5,7 +5,7 @@
 from collections import namedtuple
 from typing import Any, List
 
-from egg import EGraph, Rewrite, vars
+from snake_egg.snake_egg import EGraph, Rewrite, vars
 
 # Operations
 And = namedtuple("And", "x y")
